reorganize go files
